Inline strike/carryover/rangestrike into their single call sites
PR #83 extracted these as standalone exported functions from Battle,
justified by call-site counts that included test files. Counting only
production call sites, each of strike, carryover, and rangestrike had
exactly one caller (all in game/battle.ts), with small bodies - no
size/complexity reason to keep them standalone. performAttack keeps
its free-function shape since it has two genuine internal callers.

engine.test.ts drives the inlined behavior through gameBattle's
mutators (the same path game/battle.ts uses) instead of calling the
removed functions directly, keeping their assertions unchanged.
