+++ src/game/models/battle.ts
@@ -29,5 +29,5 @@ export {
 } from "./battle/strike"
 export type { ActiveStrikeHit, IActiveStrike, RangestrikeTarget, Strike } from "./battle/strike"
 
-export { Battle, carryover, nextPhase, rangestrike, strike } from "./battle/engine"
+export { Battle, nextPhase, performAttack } from "./battle/engine"
 export type { BattleSide } from "./battle/engine"
